add argparse overwrite option

+++ calibrate.py
@@ -12,6 +12,14 @@ from Validation import Validation
 from sanityPreCheck import RunPreCheck, RunCalibCheck, RunPreSubmitTest
 import accessories as acc
 
+
+# ---- PARSE INPUT ARGS----
+
+
+parser = argparse.ArgumentParser()
+parser.add_argument("--overwrite", action="store_true", help="Overwrite exixting directory")   
+
+
 # ----- log -----
 suffix = datetime.datetime.now().strftime("%Y-%m-%d_%H%M%S")
 logfile = 'testing.log'
@@ -28,19 +36,26 @@ logger = logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
 
 
+
 # -----  main ------
 setupfile = 'setup.yaml'
 calibrationfile = 'calib_params.tbl'
 
-#if not RunPreCheck(setupfile).run_all(): sys.exit()
-#if not RunCalibCheck(setupfile).run_all(): sys.exit()
+if not RunPreCheck(setupfile).run_all(): sys.exit()
+if not RunCalibCheck(setupfile).run_all(): sys.exit()
 
 # create the setup instance
 setup = SetMeUp(setupfile)
 
 # remove dir if it exists...
+
 if setup.parent_directory.exists():
-    shutil.rmtree(setup.parent_directory, ignore_errors=True)
+    if args.overwrite == True:
+        shutil.rmtree(setup.parent_directory, ignore_errors=True)
+    else: 
+        logger.error("directory already exists. exiting. use overwrite flag to delete")
+        sys.exit()
+    
 
 # get the current directory 
 cwd = os.getcwd()
@@ -48,20 +63,19 @@ cwd = os.getcwd()
 # Calibrate
 calib = Calibration(setupfile)
 calib.PrepareCalibration()
-#calib.AdjustCalibTable()
+calib.AdjustCalibTable()
 
 #logger.info(calib.df)
-#calib()
-calib.OneLoop()
+calib()
 
 
-#logger.info(calib.failed_iterations)
+logger.info(calib.failed_iterations)
 # make sure we are in the parent directory...
-#os.chdir(cwd) 
+os.chdir(cwd) 
 
 # Validate
 #---------
-#valid = Validation(setupfile)
-#valid.PrepareValidation()
-#valid.run_validation()
-#valid.aggregate_results()
+valid = Validation(setupfile)
+valid.PrepareValidation()
+valid.run_validation()
+valid.aggregate_results()
